using a visitor to decouple the orchestrations discovery from analysis (#300)
I realized that `OrchestrationAnalyzer` base class was performing two responsibilities:

1. Discovering orchestrations
2. Recursively finding methods invocations

While this was ok for the first 3 concrete implementations (DateTime, Guid and Delay analyzers), I couldn't properly use the base class to inspect discovered orchestrations and inspect their method signatures (when looking for Cancellation Tokens, checking bindings, etc).
The discovery behavior was tightly coupled to the recursive method invocation finding. 

In order to decouple it, I propose we introduce `OrchestrationVisitor`, a class that will contain the logic to be executed after an orchestration is found. `MethodProbeOrchestrationVisitor` represents the existent recursive method invocation behavior it had before.

This abstraction greatly simplified the code, allowing us to even remove the usage of concurrent collections and compilation end actions.