Add health-check timeout to gRPC workitem streaming (#390)
This commit adds a timeout to the gRPC stream used to communicate
with the backend. This was done because the backend could restart
and drop the connection and the worker would not know. This causes
the worker to hang and not receive any new work items. The fix is
to reset the connection if a long enough period of time has passed
between receiving anything on the stream.

Signed-off-by: halspang <halspang@microsoft.com>