Introduce versioning to the DurableTaskClient
This change adds a DefaultVersion to the DurableTaskClient builder
options. When no other version is set, which is currently always the
case, this value is used as the version of any new orchestration
started. The value is then passed down to workers via the
TaskOrchestrationContext.

This allows users to specify a version during their app setup and
then key off of the version during orchestration. Using conditional
logic, changes can then be made to the orchestration without harming
in-progress orchestrations.

Signed-off-by: halspang <halspang@microsoft.com>