Add version to TaskName
This commit adds an explict set for Version in TaskName. It also
includes a flag that is set so other code can determine if the
TaskName is intended to be versioned. Finally, it updates the
Version property in the TaskOrchestrationContext as that was always
set to empty instead of being able to be used.

Signed-off-by: halspang <halspang@microsoft.com>