Separate version from TaskName (#416)
Separate version from TaskName 

This commit adds an explict new struct for version, the `TaskVersion`.
This is used in the `TaskOptions` when starting a new orchestration instead
of exposing the `Version` field in the `TaskName`. This was done to maintain
backwards compat.

This allows for orchestrations (and sub-orchestrations) to be started with
a specific version instead of just using the `DefaultVersion` in the versioning
options.

Signed-off-by: halspang <halspang@microsoft.com>