Add ability to filter orchestrations at worker (#443)
Add ability to filter orchestrations at worker

This change allows a user to implement an interface that allows
for programmatic orchestration filtering. As it is an interface,
it can be plugged into DI. Returning true allows the
orchestration to proceed, false will block the orchestration.

In this initial commit, the filter can work off of the orchestration
name and the orchestration tags.

Signed-off-by: Hal Spang <halspang@microsoft.com>