Add version check to activities (#472)
Add version check to activities

Activities were previously not checking the versioning status of
the worker. This was because it was assumed that the orchestration
level check would satisfy the activity. However, when multiple
workers of the same version are present, the activities can be
sent to workers of the incorrect version. This change adds the
orchestration version to the activity to allow for versioning checks.

Signed-off-by: Hal Spang <halspang@microsoft.com>